LPD-34397 Move FDS data model creation to a separate component. Trigger FDS data model creation on component activation. In the future, data model will be created via object definition batch import
