+++ modules/apps/frontend-data-set/frontend-data-set-admin-api/bnd.bnd
@@ -1,0 +1,4 @@
+Bundle-Name: Liferay Frontend Data Set Admin Api
+Bundle-SymbolicName: com.liferay.frontend.data.set.admin.api
+Bundle-Version: 1.0.0
+Export-Package: com.liferay.frontend.data.set.admin.model
