+++ modules/apps/frontend-data-set/frontend-data-set-admin-api/src/main/resources/com/liferay/frontend/data/set/admin/model/packageinfo
@@ -1,0 +1,1 @@
+version 1.0.0
