+++ modules/apps/frontend-data-set/frontend-data-set-admin-web/build.gradle
@@ -11,6 +11,7 @@ dependencies {
 	compileOnly project(":apps:client-extension:client-extension-api")
 	compileOnly project(":apps:client-extension:client-extension-type-api")
 	compileOnly project(":apps:fragment:fragment-api")
+	compileOnly project(":apps:frontend-data-set:frontend-data-set-admin-api")
 	compileOnly project(":apps:frontend-data-set:frontend-data-set-api")
 	compileOnly project(":apps:frontend-taglib:frontend-taglib-react")
 	compileOnly project(":apps:info:info-api")
